requirements.txt
appdirs==1.4.3
Babel==2.3.4
cliff==2.7.0
cmd2==0.7.0
coverage==4.4.1
packaging==16.8
pbr==3.0.1
prettytable==0.7.2
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
PyYAML==3.12
requests==2.14.2
requests-mock==1.3.0
ruamel.yaml==0.14.12
six==1.10.0
stevedore==1.21.0
tosca-parser==0.8.1
